Refactored and added page changing
