+++ lib/pages/pages.dart
@@ -1,0 +1,4 @@
+library pages;
+
+export 'admin_menu_page.dart';
+export 'landing_page.dart';
